main_test.go
package main

import (
	"os"
	"testing"
)

// Mock for file operations
type MockFileOps struct {
	ReadFileFunc  func(string) ([]byte, error)
	WriteFileFunc func(string, []byte, os.FileMode) error
}

func (m *MockFileOps) ReadFile(filename string) ([]byte, error) {
	return m.ReadFileFunc(filename)
}

func (m *MockFileOps) WriteFile(filename string, data []byte, perm os.FileMode) error {
	return m.WriteFileFunc(filename, data, perm)
}

// TestSaveArticle_Success
func TestSaveArticle_Success(t *testing.T) {
	// Remove the declaration of mockFileOps

	// Assume SaveArticle is modified to accept fileOps as an argument
	err := SaveArticle(Article{Symbol: "NEW", CreatedAt: "2023-04-02", Headline: "New Article"})
	if err != nil {
		t.Errorf("SaveArticle failed: %v", err)
	}
}

// Additional tests follow a similar pattern...

// TestFetchArticles_Success
func TestFetchArticles_Success(t *testing.T) {
	articles, err := FetchArticles("AAPL", 1)
	if err != nil {
		t.Errorf("FetchArticles failed: %v", err)
	}
	if len(articles) != 1 || articles[0].Symbol != "AAPL" {
		t.Errorf("FetchArticles returned unexpected results: %+v", articles)
	}
}

// Additional tests for failure cases and filtering logic...
